add TextAreaInput component
