+++ frontend/src/components/CheckboxList/CheckboxList.css
@@ -1,0 +1,15 @@
+.form-group label {
+    display: flex;
+    padding: 0.75rem 1rem;
+    align-items: flex-start;
+    gap: 0.5rem;
+    background: #fff;
+    
+    input {
+        border: 1px solid #e7e6e4;
+        &:focus {
+            border-color: #2f6fed;
+            outline: none;
+        }
+    }
+}
