+++ frontend/src/components/TextAreaInput/TextAreaInput.css
@@ -1,0 +1,11 @@
+.form-group textarea {
+    height: 11rem;
+    padding: 0.75rem 1rem;
+    border-radius: 0.75rem;
+    border: 1px solid #e7e6e4;
+    background: #fff;
+    &:focus {
+        border-color: #2f6fed;
+        outline: none;
+    }
+}
